Running gRPC Server independently (#158)
The loading/unloading behavior of the Vulkan/OpenXR Dive layer library varies significantly across Android applications.
For example for Vulkan apps it loads - unloads - reloads the library during execution. For an Unity-based OpenXR
app that creates two instances, the behavior is the same.

The gRPC Server starts when the library loads for the first time. Then, a function prevents the library from being unloaded.
This keeps both the library and server loaded exactly once, even if apps try to reload or close them.
In this way, the Server lives for the entire lifecycle of the app.