Fix a capturing issue for sysui (#186)
In the updated change, removed the capture buffer size setting by default.
And now we can set the capture buffer size by

```
adb shell setprop debug.dive.pm4.max_buffer_size  1234
```
And clear the setting by
```
adb shell setprop debug.dive.pm4.max_buffer_size \"\"
```
This way, we can control weather we want to limit the buffer, and how large we want to limit the buffer size.

Ideally there can be a configuration box in the UI to set this, maybe we can add it in the future.