Fix crash on capture of a multi-sampled, depth/stencil image (#262)
This is a "cherry-pick" of an upstream LunarG fix into our subtree.

Capturing a program like Vulkan-samples samples/performance/msaa using
the asset file will crash. The cause is an image with samples >
VK_SAMPLE_COUNT_1_BIT and format == VK_FORMAT_D24_UNORM_S8_UINT.
VK_FORMAT_D24_UNORM_S8_UINT has two aspects (stencil and depth) so the
image is processed twice in ProcessImageMemoryWithAssetFile. The first
time, ResolveImage would fail (since it's not a color image) but the
code assumes that it will succeed; the second time (given that
assumption) hits the assert. The fix is to change the assert into a
runtime check.